% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyinput-package.R
\docType{package}
\name{shinyinput-package}
\alias{shinyinput}
\alias{shinyinput-package}
\title{shinyinput: Better-Behaved Shiny Inputs}
\description{
Create groups of inputs that mutually depend on one another, and probably other things.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shinyworks/shinyinput}
  \item \url{https://shinyworks.github.io/shinyinput/}
  \item Report bugs at \url{https://github.com/shinyworks/shinyinput/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
